refactor(transpile): only transpile one module after first build
After the first build, a change to a TS file will only transpile the
one module. Files added and deleted still use the full build. Full
builds still happen, but in another core so if there are transpile
errors the user will still be notified. Additionally, tslint runs in
another core for the individual files which have changed.