fix(serve): assign all ports dynamically (#727)
Up to now only the logger port was dynamically increased if another app
is running already. This meant that manually specifying port arguments
for server and live reload was needed when directly using the
`ionic-app-scripts server` command. The `ionic` CLI command does
already take care of this.

This patch ensures all the ports are chosen dynamically, so it's now
possible to run two applications without any manual intervention.